backend/whisper_service/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
faster-whisper>=1.0.0
httpx>=0.25.0
python-dotenv>=1.0.0
python-multipart>=0.0.6
